studentplanner/requirements.txt
Django==1.11.3
django-bootstrap4==0.0.1
django-extensions==1.9.7
django-formtools==2.0
pytz==2017.3
six==1.11.0
typing==3.6.2
